refactor: extract fetchInstalledMcpServers to shared util (DES-296)
Moves the MCP server fetching logic from claude-adapter.ts into
src/utils/mcp-server-fetcher.ts. Adds a `format` param ('claude'|'opencode')
so opencode adapter can reuse the util with `type: 'remote'` shape.
claude-adapter.ts passes format:'claude' — no behavior change.
